Gut unnecessary redux for CreateProject
