+++ src/components/App/DefaultState.ts
@@ -2,7 +2,6 @@ import { defaultState as goalTimelineState } from "components/GoalTimeline/Defau
 import { defaultState as loginState } from "components/Login/Redux/LoginReducer";
 import { defaultState as currentProjectState } from "components/Project/ProjectReduxTypes";
 import { defaultState as exportProjectState } from "components/ProjectExport/Redux/ExportProjectReduxTypes";
-import { defaultState as createProjectState } from "components/ProjectScreen/CreateProject/Redux/CreateProjectReduxTypes";
 import { defaultState as pronunciationsState } from "components/Pronunciations/Redux/PronunciationsReduxTypes";
 import { defaultState as treeViewState } from "components/TreeView/Redux/TreeViewReduxTypes";
 import { defaultState as characterInventoryState } from "goals/CharacterInventory/Redux/CharacterInventoryReducer";
@@ -15,10 +14,6 @@ export const defaultState = {
   loginState: { ...loginState },
 
   //project
-  createProjectState: {
-    ...createProjectState,
-    success: true,
-  },
   currentProjectState: { ...currentProjectState },
   exportProjectState: { ...exportProjectState },
