--- src/components/ProjectScreen/CreateProject/Redux/CreateProjectReduxTypes.ts
@@ -1,11 +1,0 @@
-export interface CreateProjectState {
-  inProgress: boolean;
-  success: boolean;
-  errorMsg: string;
-}
-
-export const defaultState: CreateProjectState = {
-  success: false,
-  inProgress: false,
-  errorMsg: "",
-};
