+++ src/rootReducer.ts
@@ -4,7 +4,6 @@ import goalsReducer from "components/GoalTimeline/Redux/GoalReducer";
 import { loginReducer } from "components/Login/Redux/LoginReducer";
 import { projectReducer } from "components/Project/ProjectReducer";
 import exportProjectReducer from "components/ProjectExport/Redux/ExportProjectReducer";
-import createProjectReducer from "components/ProjectScreen/CreateProject/Redux/CreateProjectReducer";
 import { pronunciationsReducer } from "components/Pronunciations/Redux/PronunciationsReducer";
 import { treeViewReducer } from "components/TreeView/Redux/TreeViewReducer";
 import { characterInventoryReducer } from "goals/CharacterInventory/Redux/CharacterInventoryReducer";
@@ -18,7 +17,6 @@ export const rootReducer: Reducer<StoreState> = combineReducers<StoreState>({
   loginState: loginReducer,
 
   //project
-  createProjectState: createProjectReducer,
   currentProjectState: projectReducer,
   exportProjectState: exportProjectReducer,
